requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai>=1.12.0
httpx>=0.27.0
python-dotenv==1.0.0
numpy==1.24.3
pymysql==1.1.0
google-cloud-vision==3.4.5
google-cloud-documentai==2.20.1
pandas==2.0.3
openpyxl==3.1.2
python-docx==1.1.0
python-multipart==0.0.6
